fix(entimport): 表名注解极性颠倒、空注解物化、int 列类型映射三处修复
TestMySQL 在干净 HEAD 上即失败,根因三处:

- import.go 表名注解赋值极性颠倒:实际表名与 ent 默认表名(类型名
  复数化)不一致时,注解错误地记录了默认名(复数)而非实际名——
  生成的 schema 会指向不存在的表;一致时反而设了 Table。
  修正为:不一致时记录实际名,一致时不设(序列化为 return nil,
  与官方 entimport 一致)。
- mysql.go 有符号 int 列被映射为 field.Int32,官方映射为
  field.Int;import_test.go 的 relation mock 中 age/user_card/
  node_next/user_pets/node_children/user_spouse/value 列被误转写
  为 bigint(上游为 int),一并还原。postgres 区串受波及已回退。

验证:entimport 包全套件(mysql/postgres/text)通过;
go build ./... 与 go vet 通过。

+++ gowind/pkg/sqlorm/internal/ent/entimport/import_test.go
@@ -662,10 +662,10 @@ func MockMySQLM2MTwoTypes() *schema.Schema {
 				Name: "age",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: false,
 				},
 			},
@@ -790,10 +790,10 @@ func MockMySQLM2MSameType() *schema.Schema {
 				Name: "age",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: false,
 				},
 			},
@@ -918,10 +918,10 @@ func MockMySQLM2MBidirectional() *schema.Schema {
 				Name: "age",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: false,
 				},
 			},
@@ -1051,10 +1051,10 @@ func MockMySQLO2OTwoTypes() *schema.Schema {
 				Name: "age",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: false,
 				},
 			},
@@ -1123,10 +1123,10 @@ func MockMySQLO2OTwoTypes() *schema.Schema {
 				Name: "user_card",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: true,
 				},
 			},
@@ -1212,10 +1212,10 @@ func MockMySQLO2OSameType() *schema.Schema {
 				Name: "value",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: false,
 				},
 			},
@@ -1223,10 +1223,10 @@ func MockMySQLO2OSameType() *schema.Schema {
 				Name: "node_next",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: true,
 				},
 			},
@@ -1295,10 +1295,10 @@ func MockMySQLO2OBidirectional() *schema.Schema {
 				Name: "age",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: false,
 				},
 			},
@@ -1314,10 +1314,10 @@ func MockMySQLO2OBidirectional() *schema.Schema {
 				Name: "user_spouse",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: true,
 				},
 			},
@@ -1386,10 +1386,10 @@ func MockMySQLO2MTwoTypes() *schema.Schema {
 				Name: "age",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: false,
 				},
 			},
@@ -1440,10 +1440,10 @@ func MockMySQLO2MTwoTypes() *schema.Schema {
 				Name: "user_pets",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: true,
 				},
 			},
@@ -1512,10 +1512,10 @@ func MockMySQLO2MSameType() *schema.Schema {
 				Name: "value",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: false,
 				},
 			},
@@ -1523,10 +1523,10 @@ func MockMySQLO2MSameType() *schema.Schema {
 				Name: "node_children",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: true,
 				},
 			},
@@ -1600,10 +1600,10 @@ func MockMySQLO2XOtherSideIgnored() *schema.Schema {
 				Name: "age",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: false,
 				},
 			},
@@ -1672,10 +1672,10 @@ func MockMySQLO2XOtherSideIgnored() *schema.Schema {
 				Name: "user_pets",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: true,
 				},
 			},
@@ -1778,10 +1778,10 @@ func MockMySQLM2MJoinTableOnly() *schema.Schema {
 				Name: "age",
 				Type: &schema.ColumnType{
 					Type: &schema.IntegerType{
-						T:        "bigint",
+						T:        "int",
 						Unsigned: false,
 					},
-					Raw:  "bigint",
+					Raw:  "int",
 					Null: false,
 				},
 			},
